Initial implementation of class-based OOP architecture for CodeEntropy: - Replaced legacy procedural modules with a modular, object-oriented structure - Introduced core classes: `RunManager`, `LevelManager`, `EntropyManager` with subclasses of `VibrationalEntropy`, `ConformationalEntropy`, `OrientationalEntropy` - Implemented dependency injection for improved testability and modularity - Encapsulated configuration, logging, and data management into dedicated components - Added `@property` accessors and private attributes for better encapsulation - Removed legacy function-based modules under `CodeEntropy/calculations/` - Added new entry point via `main.py` using `RunManager` orchestration
