+++ pyproject.toml
@@ -66,7 +66,7 @@ docs = [
 ]
 
 [project.scripts]
-CodeEntropy = "CodeEntropy.main_mcc:main"
+CodeEntropy = "CodeEntropy.main:main"
 
 [tool.isort]
 profile = "black"
